+++ changes.md
@@ -3,6 +3,8 @@
 #### Bugfixes
 
 * Consider `doall` side-effectful (for the `:unused-ret-vals` linter).
+* Omit the `:no-ns-form-found` linter for Eastwood config files.
+  * Closes https://github.com/jonase/eastwood/issues/406
 
 ## Changes from 0.7.1 to 0.8.1